requirements_no_dassl.txt
# Requirements for MaPLe Sentiment Analysis (No Dassl version)
# Python >= 3.8

# Core deep learning
torch>=1.13.0
torchvision>=0.14.0

# Configuration
yacs>=0.1.8

# Text processing
ftfy>=6.1.1
regex>=2022.10.31

# Utilities
tqdm>=4.64.1
Pillow>=9.3.0

# Scientific computing
numpy>=1.23.0
scipy>=1.9.0
scikit-learn>=1.1.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0

# Optional: for better performance
# ninja  # For faster compilation